refactor: add test configs, wagmi v2, separate hooks
