+++ apps/web/package.json
@@ -13,9 +13,11 @@
     "@chakra-ui/react": "^2.5.4",
     "@emotion/react": "^11.10.6",
     "@emotion/styled": "^11.10.6",
+    "@tanstack/query-core": "^5.17.15",
+    "@tanstack/react-query": "^5.17.15",
     "@web3inbox/core": "*",
     "@web3inbox/widget-react": "*",
-    "@web3modal/wagmi": "^3.0.0-beta.4",
+    "@web3modal/wagmi": "4.0.0-alpha.0",
     "framer-motion": "^10.10.0",
     "lokijs": "^1.5.12",
     "next": "^13.4.2",
@@ -24,8 +26,8 @@
     "react-hook-form": "^7.46.1",
     "react-icons": "^4.10.1",
     "usehooks-ts": "^2.9.1",
-    "viem": "^1.10.14",
-    "wagmi": "^1.4.1"
+    "viem": "2.0.10",
+    "wagmi": "2.2.1"
   },
   "devDependencies": {
     "@types/node": "18.6.4",
