+++ packages/react/package.json
@@ -12,7 +12,7 @@
     "build:cjs": "tsup src/index.tsx --format cjs --dts",
     "dev": "tsup --watch && cp src/compiled.css dist",
     "clean": "",
-    "-------": "-----------------------------------------------------------------------------------------------------",
+    "test": "vitest",
     "lint": "eslint \"**/*.ts*\""
   },
   "dependencies": {
@@ -23,13 +23,15 @@
     "react-toastify": "^9.1.3"
   },
   "devDependencies": {
-    "@web3inbox/core": "*",
+    "@testing-library/react-hooks": "^8.0.1",
     "@types/lodash.debounce": "^4.0.7",
     "@types/react": "^18.2.0",
     "@types/react-dom": "^18.2.0",
+    "@web3inbox/core": "*",
     "eslint": "^7.32.0",
     "eslint-config-custom": "*",
     "eslint-plugin-react-hooks": "^4.6.0",
+    "jsdom": "^23.2.0",
     "postcss": "^8.4.29",
     "sass": "^1.66.1",
     "sass-loader": "^13.3.2",
