+++ packages/react/src/hooks/index.ts
@@ -1,2 +1,6 @@
-export * from "./web3inboxClient";
-export * from "./notifyStateManagement";
+export * from "./useAllSubscriptions";
+export * from "./useManageSubscription";
+export * from "./useNotifications";
+export * from "./useSubscriptionScopes";
+export * from "./useW3IAccount";
+export * from "./useWeb3InboxClient";
