+++ packages/react/src/index.tsx
@@ -1,1 +1,2 @@
 export * from "./hooks";
+export * from "./utils";
