+++ packages/react/src/utils/index.ts
@@ -1,0 +1,1 @@
+export * from "./initWeb3InboxClient";
